tfprovider: Factor out protocol5 to a separate package
In preparation for supporting protocol6 as well, we'll use a multi-package
strategy instead of having all major versions together in the same package.

This doesn't change the public API in any material way, but avoids all
of the similar-but-incompatible version-specific functions trying to crowd
into the same package namespace and thereby needing obnoxious prefixes.